pyproject.toml
[project]
name = "viral-shorts-generator"
version = "1.0.0"
description = "Automated viral YouTube Shorts generator using free-tier APIs and open-source tools"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "QuantumFacts", email = "your-email@example.com"}
]
keywords = ["youtube", "shorts", "automation", "video-generation", "content-creation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "openai>=1.12.0",
    "pysubs2>=1.6.1",
    "Pillow>=10.2.0",
    "google-auth>=2.27.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.116.0",
    "tqdm>=4.66.1",
    "colorlog>=6.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/viral-shorts-generator"
Repository = "https://github.com/yourusername/viral-shorts-generator"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
